scrapy_playwright_version/requirements.txt
psycopg2-binary
python-dotenv
scrapy
scrapy-playwright
playwright-stealth
